Changed BT backend, added Beacon funtionality
Implemented Beacon functionality. Player can deploy beacon using 2, AI
will move to it and destroy it, then continue to roam.

Realized we don't want an event driven BT for this project, or at least
I don't want to make it work when the "BT 1.0" is much better suited for
this project.